refactor localisation from vic3 into a generic jomini_parser
